WIP: Simple TicTacToe game works - currently expanding to use multiple boards. Added a Global GameState manager and extracted a large portion of the logic in GameController to there. Currently 2 boards exist and work as expected.
