+++ Assets/Scripts/GameController.cs
@@ -4,33 +4,19 @@ using UnityEngine;
 using UnityEngine.UI;
 
 [System.Serializable]
-public class Player {
-    public Image panel;
-    public Text text;
-    public Button button;
-}
-
-[System.Serializable]
 public class PlayerColor {
     public Color panelColor;
     public Color textColor;
 }
 
 public class GameController : MonoBehaviour {
+    public GlobalController globalController;
     public GameObject gameOverPanel;
-    public GameObject resetButton;
-    public GameObject startInfo;
 
     public Text[] buttonList;
     public Text gameOverText;
 
-    public Player playerX;
-    public Player playerO;
-    public PlayerColor activePlayerColor;
-    public PlayerColor inactivePlayerColor;
-
     private WinningOptions winOpts;
-    private string playerSide;
     private int moveCount;
 
     void Awake() {
@@ -39,13 +25,16 @@ public class GameController : MonoBehaviour {
         gameOverPanel.SetActive(false);
         gameOverText.text = "Game Over";
         moveCount = 0;
-        resetButton.SetActive(false);
+    }
+
+    public void SetGlobalControllerRef(GlobalController global) {
+        globalController = global;
     }
 
     void SetGameControllerReferencesOnButtons() {
         foreach (Text button in buttonList)
         {
-            button.GetComponentInParent<GridSpace>().SetGameControllerRef(this);
+            button.GetComponentInParent<GridSpace>().SetGameControllerRef(this, globalController);
         }
     }
 
@@ -55,22 +44,9 @@ public class GameController : MonoBehaviour {
         if (gameIsADraw)
             gameOverText.text = "It's a draw!";
         else
-            gameOverText.text = playerSide + " wins!";
+            gameOverText.text = globalController.GetPlayerSide() + " wins!";
 
         gameOverPanel.SetActive(true);
-        resetButton.SetActive(true);
-        SetPlayerColorsInactive();
-    }
-
-    void ChangeSides ()
-    {
-        if (playerSide == "X") {
-            playerSide = "O";
-            SetPlayerColors(playerO, playerX);
-        } else {
-            playerSide = "X";
-            SetPlayerColors(playerX, playerO);
-        }
     }
 
     void SetBoardInteractable(bool active) {
@@ -81,34 +57,8 @@ public class GameController : MonoBehaviour {
         }
     }
 
-    void SetPlayerColors(Player newPlayer, Player oldPlayer) {
-        newPlayer.panel.color = activePlayerColor.panelColor;
-        newPlayer.text.color = activePlayerColor.textColor;
-
-        oldPlayer.panel.color = inactivePlayerColor.panelColor;
-        oldPlayer.text.color = inactivePlayerColor.textColor;
-    }
-
-    void StartGame() {
+    public void StartGame() {
         SetBoardInteractable(true);
-        SetPlayerButtons(false);
-        startInfo.SetActive(false);
-    }
-
-    void SetPlayerButtons(bool toggle) {
-        playerX.button.interactable = toggle;
-        playerO.button.interactable = toggle;
-    }
-
-    void SetPlayerColorsInactive() {
-        playerX.panel.color = inactivePlayerColor.panelColor;
-        playerX.text.color = inactivePlayerColor.textColor;
-        playerO.panel.color = inactivePlayerColor.panelColor;
-        playerO.text.color = inactivePlayerColor.textColor;
-    }
-
-    public string GetPlayerSide() {
-        return playerSide;
     }
 
     public void EndTurn(int buttonPressed) {
@@ -117,9 +67,9 @@ public class GameController : MonoBehaviour {
 
         if ( gameWon || gameIsADraw) {
             GameOver(gameIsADraw);
-        } else {
-            ChangeSides();
-        }
+        } 
+
+        globalController.EndTurn();
     }
 
     public bool CheckForWin(int blockIndex) {
@@ -131,7 +81,7 @@ public class GameController : MonoBehaviour {
             int winTicked = 0;
             for (int j = 0; j < currentWinningOptions[i].Length; j++)
             {
-                if (buttonList[currentWinningOptions[i][j]].text == playerSide) {
+                if (buttonList[currentWinningOptions[i][j]].text == globalController.GetPlayerSide()) {
                     winTicked++;
                 }
             }
@@ -145,22 +95,5 @@ public class GameController : MonoBehaviour {
     public void RestartGame() {
         moveCount = 0;
         gameOverPanel.SetActive(false);
-
-        resetButton.SetActive(false);
-        SetPlayerButtons(true);
-        SetPlayerColorsInactive();
-        startInfo.SetActive(true);
-    }
-
-    public void SetStartingSide(string startingSide) {
-        playerSide = startingSide;
-        
-        if (playerSide == "X") {
-            SetPlayerColors(playerX, playerO);
-        } else {
-            SetPlayerColors(playerO, playerX);
-        }
-
-        StartGame();
     }
 }
