+++ Assets/Scripts/GlobalController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99fff3d5c0beb934ea9d5c221e72c2c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
