+++ Assets/Scripts/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 14bd2a81ae14af348bdf6f06b15095ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
